requirements.txt
python-telegram-bot>=21.0
python-dotenv
httpx>=0.27.0
